Add TournamentManagement.Query project
