+++ src/Tennis.sln
@@ -25,7 +25,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TournamentManagement.Consol
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TournamentManagement.Contract", "TournamentManagement\TournamentManagement.Contract\TournamentManagement.Contract.csproj", "{7C634D92-F8CA-4C1D-9E83-C8A8441977FC}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TournamentManagement.Common", "TournamentManagement\TournamentManagement.Common\TournamentManagement.Common.csproj", "{DE88D0AC-FD0E-4971-AC74-393B2A1DA754}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TournamentManagement.Common", "TournamentManagement\TournamentManagement.Common\TournamentManagement.Common.csproj", "{DE88D0AC-FD0E-4971-AC74-393B2A1DA754}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TournamentManagement.Query", "TournamentManagement\TournamentManagement.Query\TournamentManagement.Query.csproj", "{453107E1-9E74-4FEE-867E-E3DBF73C1C3C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -157,6 +159,18 @@ Global
 		{DE88D0AC-FD0E-4971-AC74-393B2A1DA754}.Release|x64.Build.0 = Release|Any CPU
 		{DE88D0AC-FD0E-4971-AC74-393B2A1DA754}.Release|x86.ActiveCfg = Release|Any CPU
 		{DE88D0AC-FD0E-4971-AC74-393B2A1DA754}.Release|x86.Build.0 = Release|Any CPU
+		{453107E1-9E74-4FEE-867E-E3DBF73C1C3C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{453107E1-9E74-4FEE-867E-E3DBF73C1C3C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{453107E1-9E74-4FEE-867E-E3DBF73C1C3C}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{453107E1-9E74-4FEE-867E-E3DBF73C1C3C}.Debug|x64.Build.0 = Debug|Any CPU
+		{453107E1-9E74-4FEE-867E-E3DBF73C1C3C}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{453107E1-9E74-4FEE-867E-E3DBF73C1C3C}.Debug|x86.Build.0 = Debug|Any CPU
+		{453107E1-9E74-4FEE-867E-E3DBF73C1C3C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{453107E1-9E74-4FEE-867E-E3DBF73C1C3C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{453107E1-9E74-4FEE-867E-E3DBF73C1C3C}.Release|x64.ActiveCfg = Release|Any CPU
+		{453107E1-9E74-4FEE-867E-E3DBF73C1C3C}.Release|x64.Build.0 = Release|Any CPU
+		{453107E1-9E74-4FEE-867E-E3DBF73C1C3C}.Release|x86.ActiveCfg = Release|Any CPU
+		{453107E1-9E74-4FEE-867E-E3DBF73C1C3C}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -172,6 +186,7 @@ Global
 		{433B56E0-BB90-48DA-B2B4-F2EC377CD1FE} = {32CCE864-ACCA-4912-9C18-EE3FFD706644}
 		{7C634D92-F8CA-4C1D-9E83-C8A8441977FC} = {32CCE864-ACCA-4912-9C18-EE3FFD706644}
 		{DE88D0AC-FD0E-4971-AC74-393B2A1DA754} = {32CCE864-ACCA-4912-9C18-EE3FFD706644}
+		{453107E1-9E74-4FEE-867E-E3DBF73C1C3C} = {32CCE864-ACCA-4912-9C18-EE3FFD706644}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {8A0BBE81-6905-4BD2-86DB-44B6D43B4DEF}
