+++ src/TournamentManagement/TournamentManagement.WebApi/Startup.cs
@@ -19,7 +19,7 @@ using TournamentManagement.Application.Queries;
 using TournamentManagement.Application.Repository;
 using TournamentManagement.Contract;
 using TournamentManagement.Data;
-using TournamentManagement.Data.Query;
+using TournamentManagement.Query;
 using TournamentManagement.Data.Repository;
 using TournamentManagement.WebApi.Utilities;
