+++ src/TournamentManagement/TournamentManagement.WebApi/TournamentManagement.WebApi.csproj
@@ -13,6 +13,7 @@
     <ProjectReference Include="..\TournamentManagement.Application\TournamentManagement.Application.csproj" />
     <ProjectReference Include="..\TournamentManagement.Contract\TournamentManagement.Contract.csproj" />
     <ProjectReference Include="..\TournamentManagement.Data\TournamentManagement.Data.csproj" />
+    <ProjectReference Include="..\TournamentManagement.Query\TournamentManagement.Query.csproj" />
   </ItemGroup>
 
 </Project>
